+++ spotifyCreds.json
@@ -1,4 +1,4 @@
 {
-  "access_token": "BQDB8Arb-dtPYh7EQqKBz8Su1oMrbNsFC3s1Ydg7R5Nv3TTJkKxagr9CmqGNyMrfxkOwJ8ohDq2ijw_KS65z12w_dNhe733dPchtR4fjb_cVTpCXhToYPwNGBTNnUnOKcZm0R8R2ptflIzX4m4FELkKJzRmzt0LNyS12teOPEhJrVps70ZKw8BQZ",
-  "refresh_token": "AQBq6JQS0GP5C8X41Ulfj_W09Us_dN6pxrVOy5nSS0YTVecIU9-taP_o0cklKeltBcKHXgDWTaJM_j13htd2sdHIIl0tiUH_EIGypEeF98Hw6osnP907zWbGwzcJF2gy5Xy5Aw"
+  "access_token": "BQDKep_b0vD14pkqawjhOTJWSq4-h7X3InAVmvmVEGlQ-NkEWCxqZR4F5W59i9kjAvSf0W-GFTRBFj7OTUL_4rfsfYs-wsaZmQ-CR_hAG6QCoYTxAE6RKIL8v_ZNjVb3G7YmYs1WM8CJ0c7xva2cJBpucilzphBASpYyI2M",
+  "refresh_token": "AQB-NS2ext3FlitoCHUxzcEb5GlFJywPIsFW2JmkI8zPPRKbl-xf03mPT6GiQ5W7Mietqc4bxZ2Mq3-hIhAfB-O1NF3mvlI3awPJ4Mgg0DWoLnweJdn3zkeWyHPeQZGsFr0vNA"
 }
